Conflicts:
	modules/SofaBaseTopology/EdgeSetGeometryAlgorithms.inl
	modules/SofaBaseTopology/TriangleSetTopologyAlgorithms.inl
	modules/SofaBaseTopology/TriangleSetTopologyModifier.cpp
	modules/SofaEngine/TransformPosition.inl
	modules/SofaMiscMapping/DistanceMapping.inl

Former-commit-id: 65384f80bb669c485d4c2d33cbcd7578b607cf00

+++ applications/plugins/Flexible/deformationMapping/TetrahedronVolumeMapping.h
@@ -90,7 +90,7 @@ public:
         typename Hessians::iterator hessianIt = hessians.begin();
         if( d_volumePerNodes.getValue() )
         {
-            for( int i=0 ; i<v.size() ; i++ ) v[i][0] = 0;
+            for( size_t i=0 ; i<v.size() ; i++ ) v[i][0] = 0;
             SparseMatrixEigen tetraJ( m_topology->getNbTetrahedra(), x.size()*Nin );
             SparseMatrixEigen TetraToNode( v.size(), m_topology->getNbTetrahedra() );
             for (int i = 0, nbTetra = m_topology->getNbTetrahedra(); i < nbTetra; i++)
