+++ modules/SofaBaseTopology/TetrahedronSetTopologyAlgorithms.inl
@@ -67,7 +67,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideTetrahedronsWithPla
     int nbTetra=this->m_container->getNbTetrahedra();
 
     //Number of to be added points
-    int nbTobeAddedPoints=intersectedEdgeID.size()*2;
+    unsigned int nbTobeAddedPoints=(unsigned int)intersectedEdgeID.size()*2;
 
     //barycentric coodinates of to be added points
     sofa::helper::vector< sofa::helper::vector<unsigned int> > ancestors;
@@ -150,7 +150,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideTetrahedronsWithPla
 
     //tetrahedron addition
     m_modifier->addTetrahedraProcess(toBeAddedTetra);
-    m_modifier->addTetrahedraWarning(toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
+    m_modifier->addTetrahedraWarning((unsigned int)toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
 
     m_modifier->propagateTopologicalChanges();
 
@@ -170,7 +170,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideTetrahedronsWithPla
     int nbTetra=this->m_container->getNbTetrahedra();
 
     //Number of to be added points
-    int nbTobeAddedPoints=intersectedEdgeID.size()*2;
+    unsigned int nbTobeAddedPoints=(unsigned int)intersectedEdgeID.size()*2;
 
     //barycentric coodinates of to be added points
     sofa::helper::vector< sofa::helper::vector<unsigned int> > ancestors;
@@ -260,7 +260,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideTetrahedronsWithPla
 
     //tetrahedron addition
     m_modifier->addTetrahedraProcess(toBeAddedTetra);
-    m_modifier->addTetrahedraWarning(toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
+    m_modifier->addTetrahedraWarning((unsigned int)toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
 
     m_modifier->propagateTopologicalChanges();
 
@@ -1178,7 +1178,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideRestTetrahedronsWit
     int nbTetra=this->m_container->getNbTetrahedra();
 
     //Number of to be added points
-    int nbTobeAddedPoints=intersectedEdgeID.size()*2;
+    unsigned int nbTobeAddedPoints=(unsigned int)intersectedEdgeID.size()*2;
 
     //barycentric coodinates of to be added points
     sofa::helper::vector< sofa::helper::vector<unsigned int> > ancestors;
@@ -1261,7 +1261,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideRestTetrahedronsWit
 
     //tetrahedron addition
     m_modifier->addTetrahedraProcess(toBeAddedTetra);
-    m_modifier->addTetrahedraWarning(toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
+    m_modifier->addTetrahedraWarning((unsigned int)toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
 
     m_modifier->propagateTopologicalChanges();
 
@@ -1281,7 +1281,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideRestTetrahedronsWit
     int nbTetra=this->m_container->getNbTetrahedra();
 
     //Number of to be added points
-    int nbTobeAddedPoints=intersectedEdgeID.size()*2;
+    unsigned int nbTobeAddedPoints=(unsigned int)intersectedEdgeID.size()*2;
 
     //barycentric coodinates of to be added points
     sofa::helper::vector< sofa::helper::vector<unsigned int> > ancestors;
@@ -1371,7 +1371,7 @@ void TetrahedronSetTopologyAlgorithms< DataTypes >::subDivideRestTetrahedronsWit
 
     //tetrahedron addition
     m_modifier->addTetrahedraProcess(toBeAddedTetra);
-    m_modifier->addTetrahedraWarning(toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
+    m_modifier->addTetrahedraWarning((unsigned int)toBeAddedTetra.size(), (const sofa::helper::vector< Tetra >&) toBeAddedTetra, toBeAddedTetraIndex);
 
     m_modifier->propagateTopologicalChanges();
